sesje dostep

+++ server/controllers/admin.js
@@ -11,13 +11,18 @@ const {
 
 exports.deleteUser = async (req, res) => {
   const id = req.params.id;
+
   try {
-    await user.destroy({
-      where: {
-        id: id,
-      },
-    });
-    res.send({ message: "Usunięto" });
+    if (!req.session.user)
+      res.send({ message: "Sesja utracona, zaloguj się ponownie" });
+    else {
+      await user.destroy({
+        where: {
+          id: id,
+        },
+      });
+      res.send({ message: "Usunięto" });
+    }
   } catch (err) {
     res.send({ message: err.message });
   }
@@ -27,17 +32,21 @@ exports.changeUserInfo = async (req, res) => {
   const { id, changeLogin } = req.body;
 
   try {
-    await user.update(
-      {
-        login: changeLogin,
-      },
-      {
-        where: {
-          id: id,
+    if (!req.session.user)
+      res.send({ message: "Sesja utracona, zaloguj się ponownie" });
+    else {
+      await user.update(
+        {
+          login: changeLogin,
         },
-      }
-    );
-    res.send({ message: "Zmiana przeszła pomyślnie..." });
+        {
+          where: {
+            id: id,
+          },
+        }
+      );
+      res.send({ message: "Zmiana przeszła pomyślnie..." });
+    }
   } catch (err) {
     res.send({ message: err.message });
   }
@@ -46,30 +55,34 @@ exports.changeUserInfo = async (req, res) => {
 exports.createAccount = async (req, res) => {
   const { userObject } = req.body;
   try {
-    const checkLogin = await user.findOne({
-      where: {
-        login: userObject.login,
-      },
-    });
-    if (checkLogin == null) {
-      const hashed = await bcrypt.hash(userObject.password, 10);
-      const newAcc = await user.create({
-        login: userObject.login,
-        haslo: hashed,
-        isOpiekunZakl: userObject.opiekunZ,
-        isStudent: userObject.student,
-        isDyrektor: userObject.dyrektor,
-        isAdmin: userObject.admin,
-        isDziekanat: userObject.dziekanat,
-        isOpiekun: userObject.opiekunU,
+    if (!req.session.user)
+      res.send({ message: "Sesja utracona, zaloguj się ponownie" });
+    else {
+      const checkLogin = await user.findOne({
+        where: {
+          login: userObject.login,
+        },
       });
+      if (checkLogin == null) {
+        const hashed = await bcrypt.hash(userObject.password, 10);
+        const newAcc = await user.create({
+          login: userObject.login,
+          haslo: hashed,
+          isOpiekunZakl: userObject.opiekunZ,
+          isStudent: userObject.student,
+          isDyrektor: userObject.dyrektor,
+          isAdmin: userObject.admin,
+          isDziekanat: userObject.dziekanat,
+          isOpiekun: userObject.opiekunU,
+        });
 
-      res.send({
-        message: "Konto zostało pomyślnie utworzone",
-        id: newAcc.id,
-      });
-    } else {
-      res.send({ message: "Login jest już zajęty" });
+        res.send({
+          message: "Konto zostało pomyślnie utworzone",
+          id: newAcc.id,
+        });
+      } else {
+        res.send({ message: "Login jest już zajęty" });
+      }
     }
   } catch (err) {
     res.send({ message: err.message });
@@ -78,12 +91,30 @@ exports.createAccount = async (req, res) => {
 
 exports.changeRole = async (req, res) => {
   const { action, type, id } = req.body;
-  updatedStudent = await user.update({ [action]: type }, { where: { id: id } });
-  res.send(updatedStudent);
+  try {
+    if (!req.session.user)
+      res.send({ message: "Sesja utracona, zaloguj się ponownie" });
+    else {
+      updatedStudent = await user.update(
+        { [action]: type },
+        { where: { id: id } }
+      );
+      res.send(updatedStudent);
+    }
+  } catch (err) {
+    console.log(err);
+  }
 };
 
 exports.getStudents = async (req, res) => {
-  const listStudent = await user.findAll();
-
-  res.send(listStudent);
+  try {
+    if (!req.session.user)
+      res.send({ message: "Sesja utracona, zaloguj się ponownie" });
+    else {
+      const listStudent = await user.findAll();
+      res.send(listStudent);
+    }
+  } catch (err) {
+    console.log(err);
+  }
 };
