+++ server/controllers/opiekuni.js
@@ -12,41 +12,58 @@ exports.changeStatusEdit = async (req, res) => {
   const { id, opis, komentarz, status, statusOpiekuna } = req.body;
 
   try {
-    await dziennik
-      .update({ [statusOpiekuna]: status, opis: opis }, { where: { id: id } })
-      .then(async () => {
-        if (komentarz?.length) {
-          const comment = await komentarze.create({
-            dziennikId: id,
-            userId: req.session.user.id,
-            komentarz: komentarz,
-          });
+    if (!req.session.user)
+      res.send({ message: "Sesja utracona, zaloguj się ponownie" });
+    else {
+      await dziennik
+        .update({ [statusOpiekuna]: status, opis: opis }, { where: { id: id } })
+        .then(async () => {
+          if (komentarz?.length) {
+            const comment = await komentarze.create({
+              dziennikId: id,
+              userId: req.session.user.id,
+              komentarz: komentarz,
+            });
 
-          await res.send({
-            success: true,
-            status: statusOpiekuna,
-            commentId: comment.id,
-          });
-        } else
-          await res.send({
-            success: true,
-            status: statusOpiekuna,
-          });
-      });
+            await res.send({
+              success: true,
+              status: statusOpiekuna,
+              commentId: comment.id,
+            });
+          } else
+            await res.send({
+              success: true,
+              status: statusOpiekuna,
+            });
+        });
+    }
   } catch (err) {
     console.log(err);
   }
 };
 
 exports.deleteComment = async (req, res) => {
-  const id = req.params.id;
+  try {
+    if (!req.session.user)
+      res.send({ message: "Sesja utracona, zaloguj się ponownie" });
+    else {
+      const id = req.params.id;
 
-  await komentarze.destroy({ where: { id: id } }).then(res.send(id));
+      await komentarze.destroy({ where: { id: id } }).then(res.send(id));
+    }
+  } catch (err) {
+    console.log(err);
+  }
 };
 exports.downloadFile = async (req, res) => {
-  const name = req.params.name;
   try {
-    await res.download("./user_files/" + name);
+    if (!req.session.user)
+      res.send({ message: "Sesja utracona, zaloguj się ponownie" });
+    else {
+      const name = req.params.name;
+
+      await res.download("./user_files/" + name);
+    }
   } catch (err) {
     res.send(err);
   }
@@ -54,9 +71,13 @@ exports.downloadFile = async (req, res) => {
 exports.changeStatus = async (req, res) => {
   const { id, status, statusOpiekuna } = req.body;
   try {
-    await dziennik
-      .update({ [statusOpiekuna]: status }, { where: { id: id } })
-      .then(res.send({ success: true, status: statusOpiekuna }));
+    if (!req.session.user)
+      res.send({ message: "Sesja utracona, zaloguj się ponownie" });
+    else {
+      await dziennik
+        .update({ [statusOpiekuna]: status }, { where: { id: id } })
+        .then(res.send({ success: true, status: statusOpiekuna }));
+    }
   } catch (err) {
     console.log(err);
   }
@@ -66,19 +87,23 @@ exports.updateEffects = async (req, res) => {
   const { id, opis, status } = req.body;
 
   try {
-    await efektyStudent
-      .update(
-        {
-          komentarz: opis,
-          status: status,
-        },
-        {
-          where: {
-            id: id,
+    if (!req.session.user)
+      res.send({ message: "Sesja utracona, zaloguj się ponownie" });
+    else {
+      await efektyStudent
+        .update(
+          {
+            komentarz: opis,
+            status: status,
           },
-        }
-      )
-      .then(res.send({ opis: opis, status: status }));
+          {
+            where: {
+              id: id,
+            },
+          }
+        )
+        .then(res.send({ opis: opis, status: status }));
+    }
   } catch (err) {
     res.send(err);
   }
